+++ ROS#/Messages/std_msgs/ColorRGBA.cs
@@ -15,9 +15,9 @@ namespace Messages.std_msgs
 
 		public class ColorRGBA
 		{
-			public double r;
-			public double g;
-			public double b;
-			public double a;
+			public float r;
+			public float g;
+			public float b;
+			public float a;
 		}
 }
